Refactor to share AccessToken across apps
