+++ src/filing/App.jsx
@@ -6,7 +6,7 @@ import Header from './common/Header.jsx'
 import Footer from './common/Footer.jsx'
 import BrowserBlocker from './common/BrowserBlocker.jsx'
 import Loading from '../common/LoadingIcon.jsx'
-import * as AccessToken from './api/AccessToken.js'
+import * as AccessToken from '../common/api/AccessToken.js'
 import { getKeycloak, refresh, login } from './utils/keycloak.js'
 import isRedirecting from './actions/isRedirecting.js'
 import updateFilingPeriod from './actions/updateFilingPeriod.js'
